[x64] Factor out memory handling code
